% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jconfig.R
\name{JC.get}
\alias{JC.get}
\title{Config Value}
\usage{
JC.get(CFG, keys, default)
}
\arguments{
\item{CFG}{configuration object}

\item{keys}{vector of character giving the path to the desired field}

\item{default}{default value to be used in case no field is found}
}
\description{
Get the value of a given configuration field with a default in case this field doesn't exist.
}

