% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{jconfig-package}
\alias{jconfig}
\alias{jconfig-package}
\title{jconfig: Managing configuration files}
\description{
This package provides a way to manage configuration files. Files can be loaded from disk. They can be either JSON or R. Configurations can include other configuration files with clauses of the form: "include": "path/to/other/config"
}
\section{Functions}{

Users need to use 'loadJConfig', 'get.config' and 'set.config'
}

\author{
\strong{Maintainer}: Sébastien Lamy de la Chapelle \email{sebastien@jalgos.com}

}
