% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jconfig.R
\name{load.config}
\alias{load.config}
\title{Loading configs}
\usage{
load.config(file, ...)
}
\arguments{
\item{...}{Parameters to be passed to the RJSONIO fromJSON function or source}
}
\value{
A list containing the config parameters. Everything is recursively converted to a list so acess has to be done through: "[[" or jsutils::deep.list
}
\description{
Loading configs
}
\details{
Will load the config from a JSON file or a R file. Sub configs can be included with clauses of the form: "include : path/to/other/config".
}
\seealso{
set.config, get.config
}
