% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jconfig.R
\name{set.get.config}
\alias{get.config}
\alias{set.config}
\alias{set.get.config}
\title{Setting / setting global config}
\usage{
set.config(CFG)

get.config()
}
\arguments{
\item{CFG}{new function to assign}
}
\value{
The global config for get.config
}
\description{
Functions to set and access a global config
}

