% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jconfig.R
\name{loadJConfig}
\alias{loadJConfig}
\title{Loading configs}
\usage{
loadJConfig(...)
}
\arguments{
\item{...}{Parameters to be passed to the RJSONIO fromJSON function}
}
\value{
A list containing the config parameters. Everything is recursively converted to a list so acess has to be done through: "[["
}
\description{
Loading configs
}
\details{
Will load the config from a JSON file or string. Sub configs can be included with clauses of the form: "include:path/to/other/config.json".
}

